ProcessWin32.Create(     cmd: Pathname.T;     READONLY params: ARRAY  OF TEXT;     env: REF ARRAY OF TEXT := NIL;     wd: Pathname.T := NIL;     stdin, stdout, stderr: File.T := NIL)   : T RAISES {OSError.E}
has been partially fixed concerning parameter passing.

The problem is that cmd and all the parameters are concatenated into
a string and then passed to the win32 subsystem. Thus it is not
possible to use programs or parameters with pathnames that contain
spaces.

Now, cmd is always surrounded by double quotes. If the command name
begins with a ` (backquote), this is stripped and all parameters
are enclosed by double quotes, too. In this case it is not possible
to pass double quotes, as there seems to be no escape character :-(

The chosen solution is a compromise. It maintains compatibility with
the past in most cases, and offers the possibility to preserve argument
boundaries as long as no double quotes need to be passed, though with
a somewhat contrived interface. All in all it's a hack and better
solutions are welcome. The main problems for them are:

 1. The windows interface is broken as parameters cannot be passed
    transparently.
 2. The Process.Create interface should not be altered as it is
    conceptually sound and many things depend on it.

modified: libm3/src/os/WIN32/ProcessWin32.m3