Dynamically linked programs would crash on LINUXLIBC6 platform without this patch from John Polstra. Here are his comments on the patch:
The absence of the ".type" and ".size" statements in setjmpASM.s
caused the linker to get confused.  Consequently, the call to
RTThread.Save() at line 1061 of thread/POSIX/ThreadPosix.m3 went to
the wrong address.

I also changed the ".align 2" to ".align 4".  In the ELF assembler,
the argument to ".align" is the multiple of alignment.  With the
previous object file format, the argument was the log base 2 of the
multiple.  So what used to be 2**2 is now given as 4.  This isn't an
essential part of the fix, but it helps performance.

modified: m3core/src/C/LINUXLIBC6/CsetjmpASM.s