added compiler tests from PM3:
I had to make several changes to the m3makefile and extend cm3 with some
options. The regression test will only run with cm3 5.2.2. Currently lots
of tests fail, due to changed output and other things. We will have to
examine everything carefully, but I think it will be a great benefit
to have a regression test framework for the compiler.