First set of patches to implement platform independent 64bit integer types (LONGINT) in the cm3 front end:
This is a development commit, it is not complete. I'm not even sure
if I've taken the right approach at all. Don't try to use a compiler
build with this patches; it will not work as expected.

To add the type LONGINT as a separate type, the module LongInt and
LongIntExpr have been added to m3front. The scanner has been extended to
recognize LONGINT keywords and constants (integers with suffix L, e.g.
1001L). That was the easy part. Unfortunately to pass 64 bit values to
the backend, the last_chunk variable in Target.i3 had to be incremented
from Integer.size DIV ChunkSize - 1 to Int64.size DIV ChunkSize - 1,
which means that suddenly everything in the front-end is calculated
as 64 bit values. This is obviously not appropriate for Word.Ts and
ADDRESSes, and maybe for some other calucalations I haven't found yet.
The rest of the changes is concerned with trying to `fix' the front-end
calculations with respect to the correct target word size, for which
I've added Trim and Expand procedures to modules TInt and TWord.

A compiler built with these changes will now compile the core system
again without errors and warnings that were not there before, but the
resulting compiler will refuse to do something meaningful with obscure
errors, so there must be one or two things I've overlooked. It's not
really easy to find the last bugs, though, and very time consuming.

I'll continue working on this in my spare time. Of course everybody
is encouraged to try for himself, and I'd like to get comments and
opinions on what I've done. Perhaps there's a much more promising
way to add 64 bit arithmetics?

All changes regarding to this topic are on the devel_longint_branch
and won't be committed to the main trunk until they work with all
packages in the cm3 distribution. Currently only the cm3/m3-sys
collection has been branched as there are no changes elsewhere yet.