Fixes to permit use of stack unwinding exception mechanisms (e.g., for ALPHA_OSF, SOLsun/SOLgnu).  Incorporated selected simplifications from PM3 version of the gcc backend.
Successfully bootstrapped on SOLgnu with -O1.  Only known problem is for
passing of large by-value parameters that are also accessed by nested
functions (workarounds are straightforward using READONLY and copying) -- this
is a known bug in gcc-3.3.1 and should be remedied by 3.3.3 (gcc bug #12544).
Witness the behavior of the following C program, which causes an internal
compiler error:

typedef struct {
  int a;
  int f;
} A;

A *b;

void x (A a) {
  void y () {
    a.a = 0;
  }

  b = &a;
  y();
}

For -O2 there still seem to be problems (basic block reordering and sibling
call optimisations break on Solaris) -- not sure if these are problems
inherent to the cm3 backend or not.