Undo damage done by previous commits of Michael Zoet: they broke the PPC_LINUX target, and the files had already been put in unix/linux-i386 and unix/linux-ppc.
I'm not convinced by all of the changes to RTHeapDepC.c in the LINUXLIBC6
target of the m3gc-enhanced package, but I'll leave them for now.  The version
of RTHeapDepC.c in the pm3 CVS tree should be adequate.  I'll confirm this once
I test on Linux x86 with 2.6 kernel.