I have recently added experimental compiler support for incremental/generational GC to CM3, while avoiding the need for virtual memory (VM) primitives.  The changes to the compiler and run-time are *benign* (they are disabled by default), but can be turned on by flags to m3front specified by the M3_FRONT_FLAGS variable of cm3.cfg.  These flags cause the compiler to generate explicit code to synchronize appropriately with the incremental/generational garbage collector when accessing traced locations in the heap.  Note that the overhead is only for accesses to traced locations -- accessing non-traced heap locations such as ADDRESS, INTEGER, etc. has no overhead.
The support is realised as follows:

- For incremental GC, whenever a scalar REFANY (i.e., traced reference) is loaded from the heap, from a global variable, or from a VAR/READONLY parameter (which might designate a heap location) the compiler generates a check to make sure the reference is not to a grey page.  If it is, the thread will first clean the page before proceeding.  In this way, threads can never see references to old-space objects, since the only way to obtain one is by accessing a grey page.  Because this check is "eager" (i.e., we don't wait until we actually try to access the grey page) it means checks are unnecessary for scalar references held in the stacks or registers.  It also means that the garbage collector must ensure that thread stacks/registers never hold references to grey pages after synchronizing with the threads at the beginning of GC.  This is accomplished using the FinishThreadPages call in RTCollector.m3, which has been used previously for VM-synchronized incremental/generational GC on platforms (i.e., Windows) where systems calls are not atomic, so references from the stacks/registers must be kept unprotected.

- For generational GC, when a traced value (i.e., containing a reference) is written to the heap, the compiler generates a check to "dirty" (cf., "unprotect", though not necessarily using VM protection primitives) the corresponding page.

Changes to the runtime system ensure that code has been compiled with the necessary checks, otherwise disabling incremental/generational GC.  The support mechanisms build on the existing incremental/generational implementation, keeping track of the "protection" state of each page in the same way as for the VM-synchronized implementations, but without actually asking the OS to really protect the page.  Thus, this compiler-supported incremental/generational GC approach needs NO SPECIAL SUPPORT from the hardware or OS.  Of course, there is added software overhead to each reference load when using this scheme, so compiler flags have to be set in order to compile in the checks.

The motivation for all of this is a push I am making to support system threading (using pthreads) on symmetric multi-processor (SMP), multi-core (e.g., Intel's new hardware), and simultaneous multi-threaded (SMT) platforms.  On such platforms VM-synchronized incremental GC is impractical since it requires stopping all other threads while running the incremental garbage collector (since the GC must unprotect the grey pages as it cleans them, and other threads have to be prevented from accessing those pages while they are being cleaned).  The compiler-supported approach solves this by having threads synchronize with GC when accessing *traced* data in the heap.

By default the compiler will not emit the checks.  The compiler support is turned on by passing the flags IncGC and/or GenGC to m3front, using the M3_FRONT_FLAGS options of cm3.cfg.  Note that *all* compilation units must have the flags enabled for this to work, so libraries must be bootstrapped with the options enabled.  Otherwise the run-time system simply reverts to non-incremental/non-generational GC.  Also, if IncGC support is compiled in but *not* GenGC support, then the run-time system will piggy-back generational GC checks on top of incremental GC checks, by optimistically assuming that any reference loaded by a thread might be used also to write to that page (this has the effect of preserving the invariant that heap references in threads stacks/registers always refer to "unprotected" non-grey/dirty pages).

The end effect is also that incremental/generational GC no does not need all the complicated wrapper support for system calls.  Because of the invariant regarding FinishThreadPages, which assures that references on the stack or in registers can never refer to a grey page there is no need to check before passing a heap reference to a system call.  Similarly, assuming that a system call cannot write a traced reference into a heap location (highly unlikely, since the system knows nothing of allocation in the heap) there is no need to synchronize with the collector at all.  Caution is needed whenever explicit operations such as memcpy are used to copy traced state from one place to another, but such operations generated implicitly by the cm3 compiler are checked appropriately and do not need manual attention.

I would appreciate feedback on the compiler-supported incremental/generational GC.  In particular, there are a number of improvements I would like to add, but have deferred until native threading is supported, viz.:

- thread-local allocation pools: this will associate a different newPool (allocation page) to each thread and avoid the need to synchronize (RTOS.LockHeap) on every allocation; instead synchronization will only be needed in the slow-path LongAlloc code.

- thread-local caching of page "protection" state:  I'd like to avoid having to synchronize (RTOS.LockHeap) in the compiler-supported inc/gen checks when the thread knows that the page is already accessible and does not need a check.  There are various ways of doing this.  One is to maintain a bit-map with per thread recording page accessability.  Another approach is to use lock-free hardware instructions (e.g., atomic compare and swap) to tag object headers for objects that are no longer gray).  Of course, support for such lock-free operations would be needed from the compiler and/or run-time.  Currently we have no such support.

These are all things that we may want to consider in future, but for now I think we can confidently move toward native thread support while retaining the ability to do practical incremental/generational GC.

Comments welcome.