Eliminate cache for idle threads -- just let them die.  This simplifies debugging of threads using gdb since there are only as many system threads as live M3 threads.
Clean up state changes for better @M3showthread behavior -- this now works and
can be used to visualize thread execution in support of debugging.

Send SIGIO signal in Alert to thread being alerted in case it is stuck on a
select call.

[Unfortunately none of these changes appear to fix the hangs of mentor running
on Darwin -- need to see if this problem is also present on non-Darwin
(Linux/Solaris) pthreads platforms.  Assistance diagnosing the mentor problems
on Darwin would be appreciated.  For now, it appears that the thread listening
for X events is alive and responsive, but other threads appear to be stuck
waiting on mutexes.]