Fix it again :-/
The return code of do_{checkout,build,compile} must not be overwritten
by the return code of the `tee log' pipe. We could use the -o pipefaile
option of bash, but that wouldn't be very portable. It's not Bourne shell
as far as I know. So the only option is to save everything to a temporary
file and cat it after remembering the return code. I don't like this
at all :-( If anybody knows of a better solution, I'd like to hear of it.