This checkin fixes something Jay left behind.
Please check things build on platforms other than yours before making
changes.  The VAL is necessary on some targets because 0 is INTEGER whereas
result is Utypes.off_t (LONGINT on I386_DARWIN) and cannot be directly
compared with an INTEGER (result).  The VAL is how we do the conversion to
LONGINT.

On platforms where Utypes.off_t is INTEGER this code will work the same as it
is.