instead of probing  -d ./arg  or -d root/m3-sys/arg (was temporary workaround)  or -d root/arg (accidentally hit for m3cc)  or -d ./`pkgpath arg`  or -d root/`pkgpath arg` (the usual hit)
 only support:
 -f root/`pkgpath arg`/src/m3makefile

in the past at least on Windows pkgpath was a full path but
it has been relative paths a while; anyone with an old
Windows PKGS must now regenerate, ok

Automated uses within the source tree and Tinderbox always used the last form.
Interactive uses elsewhere might be broken.

The reasons this doesn't break normal uses are multiple and redundant:

 - none of the packages were immediately under the root
   (except m3cc by accident)
 - or presumably the working directory
   which presumably is either / or ~ or $cvsroot or $cvsroot/scripts
   On the off chance that wd was $cvsroot/m3-foo, then that is broken.
 - pkgpath depends on foo/src/m3makefile in populating PKGS

(Note that on Windows we don't discern file vs. directory, but we could.)