Pickles in both PM3 and CM3 have some hard-coded fingerprints for common types, but the values are not consistent about byte order.
Fix CM2 pickle-reading code to accept values of types
NULL, ROOT, UNTRACED ROOT, ADDRESS, AND REFANY,
written by PM3 pickle-writing code.  Probably,
the latter two can't even happen in a pickle, unless
somebody wrote a custom special for them.

This still won't handle a more complex type that depends
on these types, e.g., an object that contains a field of
one of them.  This latter is probably not feasable, because
it would require "decrypting" fingerprints, and the hash
algorithm for fingerprints is lossy.