Some progress adapting the the new system of static links used by gcc 4.3.0.  In many cases, m3gdb can now properly follow them and also create them, when calling or passing a nested procedure.
The main place it will fail is if a nested procedure makes no nonlocal
references to variables or parameters of an enclosing procedure, nor
does any further enclosed procedure.

If it can't find a static link, it will still fall back to the process
of nearest dynamic ancestor activation record that belongs to the
desired procedure.  This will be correct when nested procedures are
called only as procedure constants.

For older compilers everything should work, though the implementation
is greatly changed.