Replace Uexec.waitpid with new portable (C-wrapped) Uwaitpid. Thus removing its endian-dependent-ness and otherwise mostly header cloning-ness.   (even C wrappers resemble header cloning, but in a simpler more portable way)
Take care to use Uwaitpid in m3-libs/sysutils via copying,
 so that bootstrapping new sysutils against old m3core works.
 (Thus why Uwaitpid.{c,i3} deleted and replaced by Uwaitpid.quake).


This very likely fixes a bug in {I386,AMD64}_DARWIN, but
 that is not confirmed, and could be fixed otherwise.
This also very likely fixes a bug in some platforms that use linux-common or openbsd-common.
Note that as long as children processes succeed, probably no bug.
zero is zero, no matter little endian or big endian.
And likely the only important point is if the value is zero or non-zero, so again ok.


cm3root.txt: possible spec for a change to cm3 to "reduce external automation"