Remove the Sleep from [Process]Wait in sysutils, as it was removed from libm3, at least for non-user threads platforms.
This should accomplish some speedup, but not measured.

NOTE: historically m3core is built for userthreads or kernelthreads;
now sysutils is too. Ideally such "switches" are contained within
one .dll and then exported to others that adapt. In this case,
both .dlls contain the constant.

The roundabout structuring here, as well as the previous non-ideal situation,
is in the interest of letting sysutils work against old m3core, while
still getting this fix.

Another approach to consider, if it can be easily reliably achieved, is to
only do the "copying" when building against an old m3core, and go ahead
and use the usual share-by-call mechanisms when building against newer m3core.

Another approach is..what is this import_pkg quake feature?
It seems to allow sharing-by-automated-copy instead of sharing-by-call,
like static linking vs. dynamic linking.

There does seem to be a bunch of redundancy in sysutils compared to
m3core or libm3..but the compiler works with older m3core/libm3 whereas
there sort of is no such thing as an 'old' sysutils, since it was
introduced relatively recently.

Really just need to bump forward the bootstrap start??