An experimental rework of the CM3 TEXT-processing code found in m3-libs/m3core/src/text, implementing the Modula-3 required interface Text.
The types and invariants of the data structure are not changed, thus
avoiding disruption of pickles.  or anything else that depends on the
internal data structure of TEXT.

Algorithms are changed to improve performance, especially when a value
is constructed by a linear series of concatenations,(i.e.,
left-to-right or right-to-left).  Generally, Concat is slower, but
this is more than offset by gains in the accessing operations.  More
storage is allocated, but most is garbage, and retained storage ranges
from somewhat less to somewhat more.  Recursion depth is dramatically
reduced.  See the README file for more details.