Experimental support for compare_and_swap (CAS) and compare_and_set (CASP) builtin operations. CAS is defined as follows:
CAS(v, o, n): BaseType(v)

where v is an integer variable (INTEGER/LONGINT), and o and n are assignable to v,
atomically sets v to value n, so long as the previous value of v is o, returning the previous
value of v in either case.

Similarly, CASP is defined as:

CASP(v, o, n): BOOLEAN

atomically sets v to n, so long as the previous value of v is o, returning TRUE if the
comparison is successful and n was assigned to v.

I may yet generalize these to permit v to have any ordinal type.
In unsafe modules I may also permit v to be a subtype of ADDRESS.