Remove platform-specificity of dynamically allocated pthread_mutex_t and pthread_cond_t.
Replace them with UNTRACED REF ARRAY OF CHAR where
size comes from a constant variable "implemented" in C.
They were already UNTRACED REFs.
This does not add any overhead/indirection beyond what was there.


On platforms where they fit in a pointer, which is common but not universal,
it would be nice to remove the heap allocation. However the allocation
is on-demand and depends on there being a distinguished not-yet-initialized value -- NIL.
That is why I have punted on making that optimization.


Also wrap pthread_mutex_destroy to workaround bugs in HP-UX and Tru64.
(This could be done in Modula-3.)


Since we are now, in a sense, lying about the types pthread_cond_t and
pthread_mutex_t, or at least changing how they can be used (or at
least created), declare them in ThreadPThread.i3 instead of Upthread.i3.
That is, if we are going to break unlikely existant consumers of Upthread.i3,
break them more severely. That is also, basically nobody is meant
to deal with pthread in Modula-3, except for purposes of establishing
a portable Modula-3 library/interface, which is what ThreadPThread.m3 is doing.
Exposing everything to everyone from the "unix" directory seems, perhaps, wrong.
(Unless, perhaps, people are porting a bunch of C to Modula-3, just
altering syntax but leaving library calls unchanged; then it is very reasonable.)



Note that the mainstream headers are unchanged here actually.
They could be drastically cut down.
Or, wait until we have user thread support, then just stop using them.
The commands about breaking existing clients applies to many platforms, but not all, yet.


(Again I think it is worth pushing this internal stuff to a separate
interface -- PThreadC.i3 but that was undone in the past.)